+++ backend/src/main/java/com/woowacourse/ternoko/controller/CoachController.java
@@ -2,7 +2,7 @@ package com.woowacourse.ternoko.controller;
 
 import com.woowacourse.ternoko.availabledatetime.domain.AvailableDateTime;
 import com.woowacourse.ternoko.availabledatetime.dto.AvailableDateTimesResponse;
-import com.woowacourse.ternoko.config.AuthenticationPrincipal;
+import com.woowacourse.ternoko.login.domain.AuthenticationPrincipal;
 import com.woowacourse.ternoko.dto.CalendarRequest;
 import com.woowacourse.ternoko.dto.CoachResponse;
 import com.woowacourse.ternoko.dto.CoachUpdateRequest;
