+++ backend/src/main/java/com/woowacourse/ternoko/controller/CrewController.java
@@ -1,6 +1,6 @@
 package com.woowacourse.ternoko.controller;
 
-import com.woowacourse.ternoko.config.AuthenticationPrincipal;
+import com.woowacourse.ternoko.login.domain.AuthenticationPrincipal;
 import com.woowacourse.ternoko.dto.CrewResponse;
 import com.woowacourse.ternoko.dto.CrewUpdateRequest;
 import com.woowacourse.ternoko.service.CrewService;
