+++ backend/src/main/java/com/woowacourse/ternoko/login/application/AuthInterceptor.java
@@ -1,10 +1,9 @@
-package com.woowacourse.ternoko.common;
+package com.woowacourse.ternoko.login.application;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 import com.woowacourse.ternoko.common.exception.ExceptionType;
 import com.woowacourse.ternoko.common.exception.InvalidTokenException;
-import com.woowacourse.ternoko.service.AuthService;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import lombok.AllArgsConstructor;
