+++ backend/src/main/java/com/woowacourse/ternoko/login/presentation/AuthenticationPrincipalConfig.java
